core: vm: sqlvm: categorize Storage methods
Split methods of Storage into files by funtionalities.
